Some refactor/cleanup
