+++ gitlab-work-tracking/app/Cargo.toml
@@ -13,6 +13,7 @@ version = "0.1.0"
 
 [dependencies]
 anyhow = "1.0"
+board_update = {path = "../board_update"}
 clap = {version = "4.0", features = ["derive", "env", "cargo"]}
 dotenvy = "0.15"
 env_logger = "0.10"
@@ -21,5 +22,3 @@ gitlab-work = {path = "../gitlab_work"}
 itertools = "0.10"
 log = "0.4"
 nullboard-tools = {path = "../nullboard"}
-
-serde = {version = "1.0", features = ["derive"]}
