--- gitlab-work-tracking/app/src/map.rs
@@ -1,17 +1,0 @@
-// Copyright 2022-2023, Collabora, Ltd.
-//
-// SPDX-License-Identifier: BSL-1.0
-//
-// Author: Ryan Pavlik <ryan.pavlik@collabora.com>
-
-use nullboard_tools::GenericList;
-
-pub fn map_note_data_in_lists<'a, T, B, F: 'a + FnMut(T) -> B>(
-    lists: impl IntoIterator<Item = GenericList<T>> + 'a,
-    mut f: F,
-) -> impl Iterator<Item = GenericList<B>> + 'a {
-    // "move" moves f into the closure, &mut avoids moving it *out* of the closure in each call
-    let map_list = move |list: GenericList<T>| -> GenericList<B> { list.map_notes(&mut f) };
-
-    lists.into_iter().map(map_list)
-}
